Avoid that pgrep detects the present script
Use the `--ignore-ancestors` parameter of `pgrep` to try to avoid "false positives" like when the user executes (when no instance of Krusader is being running):
    ./ww -fa krusader -c krusader 
but failing (Krusader is not being executed) because the `IS_RUNNING` variable is equal to e.g. "5604 /bin/bash ./ww -fa krusader -c krusader". Problems like this one were tried to be prevented previously with the `| grep -v "$CURRENT_SCRIPT_NAME"` code that was removed recently, the `--ignore-ancestors` parameter is aimed to avoid the detection of the "ww" process (because that process is an "ancestor" of the `pgrep` process).